requirements.txt
requests==2.31.0
chromadb==0.4.22
langchain==0.1.0
streamlit==1.31.1
graphene==3.3.0
python-dotenv==1.0.0
openai==1.12.0
typing-extensions==4.9.0
pydantic==2.6.1
